backend/requirements.txt
requests>=2.31
beautifulsoup4>=4.12
pyyaml>=6.0
apscheduler>=3.10
playwright>=1.58
